#128: refactor ActiveInteractor::Model
