+++ lib/active_interactor/context/attributes.rb
@@ -114,6 +114,22 @@ module ActiveInteractor
 
       private
 
+      def _called
+        @_called ||= []
+      end
+
+      def copy_called!(context)
+        value = context.instance_variable_get('@_called') || []
+        instance_variable_set('@_called', value)
+      end
+
+      def copy_flags!(context)
+        %w[_failed _rolled_back].each do |flag|
+          value = context.instance_variable_get("@#{flag}")
+          instance_variable_set("@#{flag}", value)
+        end
+      end
+
       def merge_attribute_values(context)
         return unless context
